requirements.txt
pillow
pyautogui
opencv-python
numpy
flask